app/requirements.txt
Flask
psycopg2-binary
gunicorn
python-dotenv
Werkzeug
PyJWT
requests
chmod